requirements.txt
clip-retrieval==2.0.0
accelerate==0.19.0
diffusers==0.14.0
wandb==0.15.3
transformers==4.29.2
torchvision==0.15.2
opencv-python==4.7.0.72
gradio==3.43.2